More static_cast to satisfy MSVC static code checks.

+++ src/Vector/BLF/LogContainer.cpp
@@ -67,7 +67,7 @@ DWORD LogContainer::calculateObjectSize() const
 {
     return
         internalHeaderSize() +
-        compressedFileSize;
+        static_cast<DWORD>(compressedFileSize);
 }
 
 WORD LogContainer::internalHeaderSize() const
